Refactored log collection logic.
Moved the log collection from the 'Experiment.Run' to the 'System' classes.
This allows to track and collect multiple log file sets (one per system).
A trait LogCollection that implements basic pattern-based collection can be used for that.

The LogCollection collection trait is currently implemented by the following systems:

* HDFS1 and HDFS2;
* Yarn;
* Flink;
* Spark.

In addition, also moved 'ensureFolderIsWritable' to the 'shell' object.