Fix isRunning for Apache Flink
It used to be that peel checked if a certain flink system was running
only by checking the output of `ps`. This could lead to false positives,
when there were unrelated flink systems running. With this change a PID
directory is configured to create files containing the process ID. The
`isRunning` method now uses those PID files to check whether flink is
running.