Fix HDFS 1 format procedure
Setting the permissions of the data directory before creating it leads
to a `RuntimeException`. Furthermore was the old way of using `sed` to
modify the datanode's VERSION file not portable since it only worked
with GNU `sed` and not with BSD `sed` or all the other POSIX compliant
`sed` implementations.