Changed location of host-specific configs.
Currently, host-specific configs are located directly under the
${app.path.config} folder and the suggested way to maintain them
is via git submodules.

This, however, tends to cause more harm than good if (i) the number
of included host-specific config grows (can happen on bundles projects
with multiple collaborators), or (ii) the configs are frequently updated
as it forces extra commit that updates the submodule commit in all
dependent bundles.

To get rid of these problems we will revamp the structure as follows.

1) Host-spefcific configs reside are located under a 'hosts' subfolder of
${app.path.config}.
2) The contents of the 'hosts' folder are ignored via .gitignore.
3) The user can then do a regular git clone or create a soft link to a
common location for a specific Peel configuration.

This commit represents the changes required for point (1) of the above list.