Keep Scala experiments definitions in source form.
Currently, Scala exerpimentis defs and packaged as part of the extensions package.

This commit reworks the ApplicationContext factory to use exclusevely either an XML
or a Scala based Spring configuration based on a name convention. Scala configurations
are thereby kept in the `${app.path.config} folder and lazily recompiled.

Fixes #68.