Generate sequential ExperimentEvent IDs.
Hash based ID synthesis leads to collisions when the number of events explodes to
hundreds of thousands. As this situations occurs oft with the DstatEventExtractor
and lenghty jobs, we are switiching to an event ID generation scheme which is
handled by the EventExtractors.

With this scheme, each extractor is seeded to a unique 32-bit prefix determined
by the name of the processed file and the hash of the compainion object.

The event IDs are then sequentially generated based on this 32bit prefix.