Adds support for clusters w/o shared file system
Normally Peel expects the files of a system to be stored in a folder
shared among the nodes in the cluster. This commit allows distibution
via rsync in clusters without a shared folder.

To enable this option, one has to set the flag
`system.<system>.path.isShared = true` in the <system>'s configuration file.