Fix `start` and `stop` commands for `Yarn` system.
The current implementation explicitly starts a `resourcemanager` and (a single)
`nodemanager` on the current host.

With this change the `Yarn` system starts using the default commands for start
and stop, i.e. - `start-yarn.sh` and `stop-yarn.sh`, which will take the
current `slaves` setting into account and start `nodemanager` instances on all
hosts.